Add NatPmpManager and split CLI logic
Introduce NatPmpManager to manage natpmpc keep-alive (runs every 45s, emits portAcquired/portLost/natpmpcMissing). Add protonvpncli.cpp which moves all protonvpn CLI-related QProcess logic out of vpnmanager.cpp into a dedicated implementation file. Update VpnPage to use NatPmpManager (show/hide port row, copy port from manager, show banner if natpmpc missing, track city features to display port-forwarding notice on startup). Update MainWindow to warn that forwarded port lease lapses when app closes. Update CMakeLists to include new CLI sources and add natpmp manager headers.
