+++ src/pages/vpnpage.h
@@ -14,9 +14,8 @@
 #include <QGuiApplication>
 #include <QVersionNumber>
 #include <QHBoxLayout>
-#include <QProcess>
-#include <QRegularExpression>
 #include "../vpnmanager.h"
+#include "../cli/natpmpmanager.h"
 #include "../widgets/elidalabel.h"
 #include "../widgets/pickerbase.h"
 #include "../widgets/infobanner.h"
@@ -134,7 +133,7 @@ public:
     // Called when VpnManager has parsed a city from `protonvpn status`.
     void onStatusCityKnown(const QString& city);
     // Returns true while the natpmpc keep-alive loop is running (port forwarding active).
-    bool isPortForwardingActive() const { return m_natPmpTimer != nullptr; }
+    bool isPortForwardingActive() const { return m_natPmpManager && m_natPmpManager->isRunning(); }
 
 signals:
     void connectRequested(const QString& country, const QString& city);
@@ -171,17 +170,18 @@ private:
     int   m_checkingSpinnerFrame = 0;
     QString m_rawError;
 
-    // Port forwarding (natpmpc keep-alive loop)
-    QWidget*     m_portRow       = nullptr;
-    QLabel*      m_portLabel     = nullptr;
-    QTimer*      m_natPmpTimer   = nullptr;
-    bool         m_natPmpActive  = false;
-    int          m_forwardedPort = 0;
-    InfoBanner*  m_natpmpcBanner = nullptr;
+    // Port forwarding (NatPmpManager keep-alive loop)
+    QWidget*        m_portRow       = nullptr;
+    QLabel*         m_portLabel     = nullptr;
+    NatPmpManager*  m_natPmpManager = nullptr;
+    InfoBanner*     m_natpmpcBanner = nullptr;
 
     VpnState m_currentState = VpnState::Unknown;
     QString  m_activeCity;
+    QString  m_connectedCountryCode; // country code of the currently connected server (e.g. "US")
     QString  m_pendingStatusCity; // city from protonvpn status, applied after cities populate
+    QString  m_currentCityFeatures; // features string (e.g. "P2P") for the currently active city
+    QString  m_lastConnectedInfo;   // base info text from the last Connected state change
     bool     m_hadUnknownConnection = false;
     bool     m_stateKnown = false;
     // nullopt = citiesReady has not yet fired for the local country;
@@ -200,7 +200,7 @@ private:
     void applyFreeUserMode();
     void startNatPmpLoop();
     void stopNatPmpLoop();
-    void runNatPmp();
+    void refreshConnectedInfoLabel();
     // After populate(), try to select m_pendingStatusCity; falls back to
     void applyPendingStatusCity();
 };
